Add server-side admin REST API, fix top-users type mismatch
Route admin dashboard data through /api/admin/* endpoints instead of
calling Supabase directly from the browser, so each admin_* RPC's own
is_admin_user() check is the sole gate on that data. Also fixes
admin_get_top_users() raising "structure of query does not match
function result type": sum(real) doesn't implicitly cast to the
declared numeric column type.

+++ supabase/admin.sql
@@ -115,7 +115,7 @@ begin
     p.preferred_key,
     p.onboarding_complete,
     count(ps.id) as session_count,
-    coalesce(sum(ps.duration_sec), 0) as practice_sec,
+    coalesce(sum(ps.duration_sec), 0)::numeric as practice_sec,
     (select count(*) from public.lesson_progress lp where lp.user_id = p.id) as lessons_completed,
     u.last_sign_in_at as last_sign_in,
     p.created_at
